requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.21.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
